Move static utilities into the http handler base class.
